requirements.txt
pandas
numpy
requests
biopython
chembl_webresource_client
pydantic
scikit-learn
rdkit
seaborn
matplotlib
XGBoost
-e .